+++ packages/authentication/package.json
@@ -25,7 +25,7 @@
   "dependencies": {
     "@loopback/context": "^4.0.0-alpha.24",
     "@loopback/core": "^4.0.0-alpha.26",
-    "@loopback/openapi-specgen": "1.0.0-alpha1",
+    "@loopback/openapi-v2": "1.0.0-alpha1",
     "@loopback/rest": "^4.0.0-alpha.15",
     "passport": "^0.4.0",
     "passport-strategy": "^1.0.0"
