--- packages/openapi-specgen/src/index.ts
@@ -1,1 +1,0 @@
-export * from './decorators/controller-decorators/metadata';
