+++ packages/openapi-v2/src/index.ts
@@ -1,0 +1,1 @@
+export * from './controller-spec';
