+++ packages/openapi-v2/test/unit/controller-spec/controller-decorators/param-decorators/param-path.test.ts
