+++ packages/rest/package.json
@@ -26,7 +26,7 @@
     "@loopback/context": "^4.0.0-alpha.23",
     "@loopback/core": "^4.0.0-alpha.25",
     "@loopback/openapi-spec": "^4.0.0-alpha.17",
-    "@loopback/openapi-specgen": "^1.0.0-alpha.1",
+    "@loopback/openapi-v2": "^1.0.0-alpha.1",
     "@types/http-errors": "^1.6.1",
     "body": "^5.1.0",
     "debug": "^3.1.0",
